add api for use in other plugins or scripts
